requirements.txt
langchain>=1.0.8
langchain-community>=0.4.1
langchain-ollama>=1.0.0
langchain-chroma>=0.1.2
chromadb>=1.3.5
fastembed>=0.7.3
python-dotenv>=1.2.1
python-pptx>=0.6.21
python-docx>=1.1.0
odfpy>=1.4.1